requirements.txt
numpy
pandas
scikit-learn
jupyter
ipykernel
tqdm
openai
nltk
rouge-score
sentence-transformers
pyyaml>=6.0

# AutoDDG
git+https://github.com/VIDA-NYU/AutoDDG@main
